# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:4000/apaman198/"
    },
    {
      "url": "https://localhost:4000/apaman198/"
    }
  ],
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0-oas3",
    "title": "Montaman API Document",
    "termsOfService": "https://monta.ir/apaman",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
...
  "paths": {
...
    "/announce/{BuildingID}.{UnitID}": {
      "get": {
        "tags": [
          "announce"
        ],
        "summary": "Get the list of all announcements",
        "description": "Using this endpoint one can gives the complete list of all announcements of a building",
        "operationId": "announceSelectAll",
        "parameters": [
          {
            "name": "BuildingID",
            "in": "path",
            "description": "The id of building that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "UnitID",
            "in": "path",
            "description": "The unit of building that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementDetails"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          },
          "500": {
            "description": "server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "post": {
        "tags": [
          "announce"
        ],
        "summary": "Create a new announcement",
        "description": "Using this endpoint one can create a new announcement",
        "operationId": "addAnnouncement",
        "parameters": [
          {
            "name": "BuildingID",
            "in": "path",
            "description": "The id of building that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "UnitID",
            "in": "path",
            "description": "The unit of building that needs to be fetched.",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "body",
            "description": "Cost object that needs to be added",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AnnouncementDetails"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AnnouncementResult"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authorization Failed!!!"
          },
          "500": {
            "description": "server Error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AnnouncementDetails": {
        "type": "object",
        "properties": {
          "PeriodDetailID": {
            "type": "integer"
          },
          "CostClassID": {
            "type": "integer"
          },
          "IsDefault": {
            "type": "integer"
          },
          "CalculationHeaderIDs": {
            "type": "string"
          },
          "Title": {
            "type": "string"
          },
          "RecordNumber": {
            "type": "integer"
          },
          "LastPayDate": {
            "type": "string"
          }
        }
      },
...
      "AnnouncementResult": {
        "type": "object",
        "required": [
          "ResultCode",
          "ResultText"
        ],
        "properties": {
          "ResultCode": {
            "type": "string",
            "example": 1
          },
          "ResultText": {
            "type": "string",
            "example": "Done"
          }
        }
      },
...
    }
  }
}